use bigint as standard int parameter type for postgres queries

+++ src/test/php/unit/sandbox.php
@@ -255,7 +255,7 @@ class sandbox_unit_tests
         $db_con->set_name('source_type_by_id');
         $db_con->set_where_std(2);
         $created_sql = $db_con->select_by_set_id();
-        $expected_sql = "PREPARE source_type_by_id (int) AS 
+        $expected_sql = "PREPARE source_type_by_id (bigint) AS 
               SELECT source_type_id,  type_name 
                 FROM source_types
                WHERE source_type_id = $1;";
