+++ src/test/resources/db/log/change_log_named_by_dsp_of_wrd.sql
@@ -1,4 +1,4 @@
-PREPARE change_log_named_by_dsp_of_wrd (int,int) AS
+PREPARE change_log_named_by_dsp_of_wrd (bigint,bigint) AS
     SELECT s.change_id,
            s.user_id,
            s.change_time,
