+++ src/test/resources/db/phrase/phrase_group_word_link_by_id.sql
@@ -1,4 +1,4 @@
-PREPARE group_word_link_by_id (int) AS
+PREPARE group_word_link_by_id (bigint) AS
     SELECT group_word_link_id,
            group_id,
            word_id
