+++ src/test/resources/db/result/result_by_usr_cfg.sql
@@ -1,4 +1,4 @@
-PREPARE result_by_usr_cfg (int, int) AS
+PREPARE result_by_usr_cfg (bigint, bigint) AS
     SELECT group_id
       FROM user_results
      WHERE group_id = $1
