+++ src/test/resources/db/result/result_list_by_source_group_id.sql
@@ -1,4 +1,4 @@
-PREPARE result_list_by_source_group_id (int) AS
+PREPARE result_list_by_source_group_id (bigint) AS
     SELECT group_id,
            formula_id,
            user_id,
